Use float coords for vvad boxes; add type ignore
Change vvad track box typing and values from ints to floats to match vvad expectations: vvad_track_boxes now holds Tuple[int, Tuple[float, float, float, float]] and appended box coords are cast with float(...). Also add a "# type: ignore" to the MediaPipe FaceMesh instantiation to silence typing errors from the mediapipe package. These are small type/annotation fixes and do not change runtime behavior beyond ensuring correct types for downstream consumers and the type checker.

+++ src/om1_vlm/anonymizationSys/face_recog_stream/face_tracker.py
@@ -460,7 +460,7 @@ class FaceTracker:
         self._active_ids = set()
         results: List[TrackResult] = []
         need_recog: List[Tuple[int, np.ndarray, Optional[np.ndarray], float]] = []
-        vvad_track_boxes: List[Tuple[int, Tuple[int, int, int, int]]] = []
+        vvad_track_boxes: List[Tuple[int, Tuple[float, float, float, float]]] = []
 
         for track in tracks:
             x1, y1, x2, y2 = map(int, track[:4])
@@ -475,7 +475,9 @@ class FaceTracker:
             self._active_ids.add(track_id)
 
             if self.vvad is not None:
-                vvad_track_boxes.append((track_id, (x1, y1, x2, y2)))
+                vvad_track_boxes.append(
+                    (track_id, (float(x1), float(y1), float(x2), float(y2)))
+                )
 
             # Get or create identity state
             if track_id not in self._identities:
